backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
beautifulsoup4
requests
langchain-groq
langchain
python-dotenv
pydantic